Fix parsing bugs, allow open file to pick multiple files, minor optimizations, refactoring
- Fix: index-out-of-range in Parser.TryParseSections (// skip indented)
- Fix: Despite the spec calling for ints, floating points are allowed
- Fix: Shape.NumObjects was not correct after clearing, make a read-property
- Optimization: Only update shape properties when shape is complete
- Optimization: Do all analysis steps after closing file handles

+++ OsuDiffCalc/FileProcessor/Beatmap.cs
@@ -66,7 +66,7 @@
 		}
 
 		public TimingPoint GetTiming(HitObject obj, bool startTime = true) {
-			int time = startTime ? obj.StartTime : obj.EndTime;
+			double time = startTime ? obj.StartTime : obj.EndTime;
 			int maxIndex = -1;
 			int numTimingPoints = TimingPoints.Count;
 			for (int i = 0; i < numTimingPoints; i++) {
